mcp-demo/requirements.txt
fastmcp==2.10.2
streamlit==1.46.1
google-generativeai==0.8.5